feat: implement 64 bit filters

+++ trailsense-edge/src/probes/fingerprint_store.rs
@@ -4,12 +4,12 @@ use core::cell::RefCell;
 use embassy_sync::blocking_mutex::{Mutex, raw::CriticalSectionRawMutex};
 use heapless::Vec as HeaplessVec;
 
-const MAX_FINGERPRINTS: usize = 2048;
+const MAX_FINGERPRINTS: usize = 2048; // TODO: This might be needed to change to 512 probably (test max size etc).
 
-static FINGERPRINTS: Mutex<CriticalSectionRawMutex, RefCell<HeaplessVec<u16, MAX_FINGERPRINTS>>> =
+static FINGERPRINTS: Mutex<CriticalSectionRawMutex, RefCell<HeaplessVec<u64, MAX_FINGERPRINTS>>> =
     Mutex::new(RefCell::new(HeaplessVec::new()));
 
-pub fn push(fingerprint: u16) -> bool {
+pub fn push(fingerprint: u64) -> bool {
     FINGERPRINTS.lock(|v| v.borrow_mut().push(fingerprint).is_ok())
 }
 
@@ -17,6 +17,6 @@ pub fn drain() {
     FINGERPRINTS.lock(|v| v.borrow_mut().clear());
 }
 
-pub fn snapshot() -> Vec<u16> {
+pub fn snapshot() -> Vec<u64> {
     FINGERPRINTS.lock(|v| v.borrow().iter().copied().collect())
 }
